Changed namespace of SyncLevel and Depth
