+++ src/main/kotlin/at/bitfire/dav4jvm/property/webdav/Depth.kt
@@ -1,4 +1,4 @@
-package at.bitfire.dav4jvm.property.push
+package at.bitfire.dav4jvm.property.webdav
 
 import at.bitfire.dav4jvm.Property
 import at.bitfire.dav4jvm.PropertyFactory
@@ -6,19 +6,19 @@ import at.bitfire.dav4jvm.XmlReader
 import org.xmlpull.v1.XmlPullParser
 
 /**
- * Represents a [NS_WEBDAV_PUSH]`:sync-level` property.
+ * Represents a [NS_WEBDAV]`:depth` property.
  *
  * Experimental! See https://github.com/bitfireAT/webdav-push/
  */
-data class SyncLevel(
+data class Depth(
     /** May be `0`, `1` or [Int.MAX_VALUE] (infinite). */
-    val level: Int? = null
+    val depth: Int? = null
 ): Property {
 
     companion object {
 
         @JvmField
-        val NAME = Property.Name(NS_WEBDAV_PUSH, "sync-level")
+        val NAME = Property.Name(NS_WEBDAV, "depth")
 
     }
 
@@ -27,10 +27,10 @@ data class SyncLevel(
 
         override fun getName() = NAME
 
-        override fun create(parser: XmlPullParser): SyncLevel {
+        override fun create(parser: XmlPullParser): Depth {
             val text = XmlReader(parser).readText()
             val level = if (text == "infinite") Int.MAX_VALUE else text?.toIntOrNull()
-            return SyncLevel(level)
+            return Depth(level)
         }
 
     }
