+++ src/main/kotlin/at/bitfire/dav4jvm/property/webdav/SyncLevel.kt
@@ -1,4 +1,4 @@
-package at.bitfire.dav4jvm.property.push
+package at.bitfire.dav4jvm.property.webdav
 
 import at.bitfire.dav4jvm.Property
 import at.bitfire.dav4jvm.PropertyFactory
@@ -6,19 +6,19 @@ import at.bitfire.dav4jvm.XmlReader
 import org.xmlpull.v1.XmlPullParser
 
 /**
- * Represents a [NS_WEBDAV_PUSH]`:depth` property.
+ * Represents a [NS_WEBDAV]`:sync-level` property.
  *
  * Experimental! See https://github.com/bitfireAT/webdav-push/
  */
-data class Depth(
+data class SyncLevel(
     /** May be `0`, `1` or [Int.MAX_VALUE] (infinite). */
-    val depth: Int? = null
+    val level: Int? = null
 ): Property {
 
     companion object {
 
         @JvmField
-        val NAME = Property.Name(NS_WEBDAV_PUSH, "depth")
+        val NAME = Property.Name(NS_WEBDAV, "sync-level")
 
     }
 
@@ -27,10 +27,10 @@ data class Depth(
 
         override fun getName() = NAME
 
-        override fun create(parser: XmlPullParser): Depth {
+        override fun create(parser: XmlPullParser): SyncLevel {
             val text = XmlReader(parser).readText()
             val level = if (text == "infinite") Int.MAX_VALUE else text?.toIntOrNull()
-            return Depth(level)
+            return SyncLevel(level)
         }
 
     }
